refactor: cactus to handle expect exception
In the test 'rl_memory_cannot_be_accessed_in_s', tftf
shares memory to the SP that it changed to realm PAS.

The access from the SP should result in an exception.
Cactus is to receive the indication it should expect an
exception on memory access via the arguments of the
'cactus_mem_send_cmd' the request. This is used to avoid
the reads to check memory, which could return in an error
response, which would make the test falsely fail.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ie220d4bab04c052a73e8b04c8667c9973b95a9e0