test: add tests for self-hosted debug
This patch adds tests for self-hosted debug.

One test verifies that the number of breakpoints/watchpoints
available in Realm is the same as what was requested by the Host.

The other test verifies that entering Realm does not corrupt
the values written to self-hosted debug registers by NS world, and
vice versa (i.e. that the registers are saved and restored correctly).

Signed-off-by: Chuyue Luo <Chuyue.Luo@arm.com>
Change-Id: I0d743c29215dd69302bab617a95c9fa507faec54