test(memory share): retrieve request from VM
The SPMC supports the case in which the hypervisor doesn't
track the operation. This is usable for multiple borrower operations,
with a VM. In this case, hypervisor/OS Kernel shall forward the
requests to retrieve memory into the SPMC. This patch adds such
case, as we lost this coverage in the Hafnium tests.

Change-Id: Ia217ababd6e4ae0a82bed781b6f217a52a0909e0
Signed-off-by: J-Alves <joao.alves@arm.com>