test(memory share): hypervisor's rx is realm on retrieve
The test validates that the SPMC is able to recover from a trapped
access to a realm RX buffer during the retrieve response. The
call to FFA_MEMORY_RETRIEVE_REQ should return FFA_ERROR_ABORTED.
At the end of the operation, the sender shall still be able to reclaim
as the retrieve request should have failed to update any of the structures.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I5d0f2a3e5716a3ed27dda4b6592682c34841ed63