refactor: refactor `get_ffa_feature_test_target`
Change `get_ffa_feature_test_target` to return a `size_t` instead of an
`unsigned int`, because `size_t` is the return type of operators like
`sizeof()`.

Change `get_ffa_feature_test_target` to require its argument to be
non-null (and assert that it is). This function is only used for getting
the array of features to test, so there is no use case where passing a
non-null pointer would make sense.

Signed-off-by: Karl Meakin <karl.meakin@arm.com>
Change-Id: I33597f1a2f7681eda59ece08062e48c28752c111