refactor: clarify which kind of exceptions it catches
The function that was called "undef_injection_handler" doesn't just
catch undef injections. It also catches traps to EL2 due to registers
not being present. Both cases have the same EC value, so it is
impossible to distinguish between them.

This patch edits variable names and adds a comment to clarify this.

Change-Id: Ie7405d7611afc1d2ff2207cfa4a08de3cbc9dff7
Signed-off-by: Charlie Bareham <charlie.bareham@arm.com>
Signed-off-by: Jayanth Dodderi Chidanand <jayanthdodderi.chidanand@arm.com>