feat(realm): set number of num_bps and num_wps
As per RMM Specification number of breakpoints and
watchpoints passed to RMI_REALM_CREATE cannot be 0.
These values are passed to host_prepare_realm_payload()
in feature_flag parameter fields which are set to 0
by default by callers to this function.
This patch modifies the logic for setting num_bps
and num_wps Realm parameters to avoid 0 values.

Change-Id: Ib5420db959866620005c404c494c4ec1904b010c
Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>