fix(realm): fix realm initialisation code
The curent code in 'realm_entrypoint' relies
on primary VCPU to perform initialisation and
symbols relocation. This makes impossible to
run any REC<n> before REC<0> runs, because
REC<n> will be detected as a secondary VCPU
not required to perform the initialisation
which will cause the crash.
This patch fixes this issue by introducing
'cold_boot_flag' variable which is updated
by the first VCPU which runs 'realm_entrypoint'
and performs initialisation.

Change-Id: I0161f7132f64423cff646db74f95753aa9a5d073
Signed-off-by: AlexeiFedorov <Alexei.Fedorov@arm.com>