refactor: hypervisor retrieve request helpers
Factored out of hypervisor_retrieve_request the looping
part, for retrieving the fragments of a fragmented
retrieve request. This is to aid testing hypervisor retrieve
request, when it faults in the middle of the operation.

Also added two helpers to access the size of a fragment
and the total size for both ABIs:
- FFA_MEM_RETRIEVE_RESP.
- FFA_MEM_FRAG_TX.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I49d12d69eff8c132b0a29954772dd6634f590f88