test(SMCCC): test SMCCC_ARCH_FEATURE_AVAILABILITY
This test calls the function with each valid argument and checks that
every bit is set if its relevant feature is present in the system. It
also fails the test if any set bit in the return value has not been
checked. This should serve as a reminder to update this test for every
new feature that is implemented.

Only feature that tfa supports are tested with the expectation that new
ones will be added in the future.

Co-developed-by: Charlie Bareham <charlie.bareham@arm.com>
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I801326a49810bb76bfc3b9d06780d416dcc32a40