feat(smc): support SMC calls with no ret vals in x8
The current implementation of the SMC library assumes that x8
contains the address of an smc_ret_values structure. Although
this can be convenient, it prevents from using registers x8-x19
as arguments, since it assumes there is an address in x8.

This patch implements an alternative API for SMC calls, allowing
the use of registers x1-x19 as input arguments and the use of
registers x0-x19 as output arguments without assuming a pointer
in x8.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Change-Id: I0016116b8d2ee4ef5aac9473f31e38434cda4943