fix(ls64): fix build with LOG_LEVEL=50 and improve test
When compiling the LS64 test with LOG_LEVEL=50, the compiler will
generate a warning about a mismatching printf format specifier:

Fix this by dropping the "ll" prefix, since the variable to be printed
is a mere integer.
Also use the opportunity to improve the test coverage a bit, and use
more complex bit patterns for the write-read-back test. This helps to
spot corruptions in the values.
Also LS64_ARRAYSIZE does not need to be hardcoded and in a header file,
since it's only used privately by the LS64 test code. Explain how the
value of "8" comes about while at it.

Change-Id: I889e614a4158c5fa5d3b9f5b056d8d10f2681755
Signed-off-by: Andre Przywara <andre.przywara@arm.com>