feat(spm): add STMM cactus partition
Add another instance of cactus S-EL1 partition, using
StMM manifest.

This is to validate HOB generation at build-time.

The manifest contains memory region nodes using the
same node names as in the StMM partition manifest,
but changed the ranges to align with memory map of
base FVP. The device region nodes have been dropped,
as they didn't affect the creation of the HOB list.

Defined the UUID for the partition in the test code,
and refactored slightly some of the code paths,
so it has an RXTX buffer and is able to do the basic
FF-A setup tests.

Also added the cactus-stmm node to the sp_layout
file.

Signed-off-by: Kathleen Capella <kathleen.capella@arm.com>
Change-Id: I05971fc9d63f03bd7ea43b3bcaba5b6362a44ca0
Signed-off-by: J-Alves <joao.alves@arm.com>