test: modify the self-hosted debug tests for debugging
This patch has an additional test case to test the emulation
logic in RMM. The test assumes that even though the hardware
supports 'X' number of Bp's or WP's, the NS-Host only requests
the minimum number of BP's = 2 during the realm creation.

The emulation logic in RMM creates the realm with the number
of BP's requested by the Host, irrespective of the number of
BP's supported by hardware.

Signed-off-by: Sona Mathew <sonarebecca.mathew@arm.com>
Change-Id: I7c700b8ee5f871a8f94f02d105422a66bd867c98