feat(mte): Specify and map MTE memory region
Specify a new ELF file section for MTE memory in the TFTF linker script.
This section can be used by tests (most likely MTE specific tests) to
specify memory that needs to be mapped with the tagged memory attribute.
We also add this memory to the initial mappings at platform boot.

Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>
Change-Id: I30581201bd9cede6bf3da0fe9bfcc17cc5b4566d