fix(smccc): availability test: add two features and remove others
The SMCCC_ARCH_FEATURE_AVAILABILITY test was not up-to-date and was
missing two features: FEAT_TWED and FEAT_PMUV3P9. Connect the SCR_EL3
and MDCR_EL3 bits to their corresponding ID register fields, so that
they can be tested.

At the same time there are two tests which are not really meaningful:
- FEAT_RNG_TRAP does not require the SCR_EL3.TRNDR bit to be set, as a
  platform would only set this bit if there is a handler returning
  random values, for instance to filter, limit, or replace values from
  the RNDR instruction. Drop the test from the list.
- The SCR_EL3.TERR bit enables trapping accesses to certain RAS
  registers, so it's not mandatory to have this bit set for FEAT_RAS.
  Drop the test from the list.

This fixes the tf-a-tests run on an FVP with ARMv9.4 enabled, which was
reporting the following issues before:
=====================================
> Executing 'SMCCC_ARCH_FEATURE_AVAILABILITY test'
  TEST COMPLETE                                                 Failed
is_feat_rng_trap_present says feature is supported but SCR_TRNDR_BIT was
	not set!
is_feat_ras_present says feature is supported but SCR_TERR_BIT was not
	set!
SCR_EL3 still has values set: 0x20000000. Test needs to be updated
MDCR_EL3 still has values set: 0x80. Test needs to be updated
=====================================

Change-Id: I73a0d240b2cd1a16e1c64d3d66ee30e658c9c946
Signed-off-by: Andre Przywara <andre.przywara@arm.com>