perf(tftf): don't use waitms() for events testing
Waiting for wall time on FVP takes longer than the desired time. As a
result this test takes a seemingly very long (over 30 seconds!) and is
sometimes indistinguishable from hanging. Additionally, it is a tftf
self test so the chances of it ever going wrong are practically zero.

Decouple the test from time and use more primitive synchronisation to
maintain the same function. Since the primary waits for all secondaries
to enter the test, we can be fairly certain that at least one has
proceeded a little further and is waiting for an event by the time the
primary has caught up and we don't need the delay.  Then the other
delays can be removed by having the secondaries wait on non-event API
event before waiting on the event.

Change-Id: I958595b93030596633bbd08b2369ecdffe34d6d3
Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>