test: exclude Boot requirement tests for Corstone-1000
A test compares the value of the Generic Timer register CNTFRQ visible in
two frames CNTBaseN and CNTCTLBase that are linked in Armv8-A and reflect
the same value.

An issue in Corstone-1000 (errata 2142118) makes the CNTFRQ views
inconsistents and the test failing. There is no workaround and then
the test is skipped.

Errata: https://developer.arm.com/documentation/sden2142076/0002/?lang=en
Change-Id: Id15011d37e3221c57f29f337365522fe23a3eb7e
Signed-off-by: Christophe Thiblot <christophe.thiblot@arm.com>