feat(amd): add test for node EEMI APIs
Add test for the node-related EEMI APIs within the TF-A test
framework. The purpose of these tests is to validate the
functionality and reliability of various node operations.

Test cover the following EEMI API functionalities:
- Retrieving node status
- Setting node requirements
- Setting maximum latency
- Releasing an already released node
- Requesting an already requested node

Note: This initial test does not cover all the node ids, as the
purpose of the test is to validate the node EEMI APIs interface.
The HW design must include USB and RTC device nodes in order to pass
this test, otherwise the test will fail.

Change-Id: I1883abb20ae7e602530e504f3c39b20d02f57004
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>