feat(amd): add test for clock EEMI APIs
Add test for the clock-related EEMI APIs within the TF-A test
framework. The purpose of these tests is to validate the
functionality and reliability of various clock operations.

Test cover the following EEMI APIs:
- xpm_clock_get_status
- xpm_clock_enable
- xpm_clock_disable
- xpm_clock_set_parent
- xpm_clock_get_parent
- xpm_clock_set_divider
- xpm_clock_get_divider

Note: This initial test does not cover all the clock ids, as the
purpose of the test is to validate the clock EEMI APIs interface.
The HW design must have QSPI nad GEM clocks enabled to pass this
test, otherwise the test will fail.

Change-Id: Ie0d7dc844d4e58d1000a9fd6d669528103b097c7
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>