feat(amd): add test for force powerdown EEMI API
Add tests for the EEMI APIs that handle force power down and request
wakeup within the TF-A test framework. The force power down API is
used to terminate an unresponsive subsystem and automatically release
its resources. The request wakeup API is used to power up a CPU node
within the same PU or to power up another PU.

Change-Id: I4fa6da87a3044c52834ab34621141cce5e136948
Signed-off-by: Madhav Bhatt <madhav.bhatt@amd.com>