refactor(sve): convert flag handling macros to functions
The code to save the value of CPTR_EL2 or CPACR_EL1 and manipulate the
SVE trap bits is implemented as a CPP macro, which makes it harder to
read and extend.
We need to enhance the functionality soon, so convert the macro into a
static C function. This also gives the compiler the opportunity to
optimise the code better.

Change-Id: I05fbc8a50ad8a38ed85cfb8f04f37ed338ae0fb4
Signed-off-by: Andre Przywara <andre.przywara@arm.com>