fix(sve): observe HCR_EL2.E2H bit when using CPTR_EL2
When running in EL2, the HCR_EL2.E2H bit controls the use of the VHE
host mode, which among other things changes the layout of the CPTR_EL2
register, to become compatible with the CPACR_EL1 register.

Take the current setting of the E2H bit into account whenever accessing
the CPTR_EL2 register, to guarantee proper operation regardless of the
E2H bit being 0 or 1.

Change-Id: I305e2f724b5dec0c17a9f822c472008e4bbdf459
Signed-off-by: Andre Przywara <andre.przywara@arm.com>